Move Block.Bulder.inline to analysis package
