+++ test/jdk/java/lang/reflect/code/linq/Queryable.java
@@ -22,6 +22,7 @@
  */
 
 import jdk.incubator.code.Op;
+import jdk.incubator.code.analysis.Inliner;
 import jdk.incubator.code.dialect.java.ClassType;
 import jdk.incubator.code.dialect.java.JavaOp;
 import jdk.incubator.code.dialect.java.MethodRef;
@@ -63,7 +64,7 @@ public interface Queryable<T> {
         JavaType queryableType = parameterized(Queryable.TYPE, elementType);
         FuncOp nextQueryExpression = func("query",
                 functionType(queryableType, queryExpression.invokableType().parameterTypes()))
-                .body(b -> b.inline(queryExpression, b.parameters(), (block, query) -> {
+                .body(b -> Inliner.inline(b, queryExpression, b.parameters(), (block, query) -> {
                     Op.Result fi = block.op(lambdaOp);
 
                     MethodRef md = method(Queryable.TYPE, methodName,
@@ -94,7 +95,7 @@ public interface Queryable<T> {
         JavaType queryResultType = parameterized(QueryResult.TYPE, resultType);
         FuncOp queryResultExpression = func("queryResult",
                 functionType(queryResultType, queryExpression.invokableType().parameterTypes()))
-                .body(b -> b.inline(queryExpression, b.parameters(), (block, query) -> {
+                .body(b -> Inliner.inline(b, queryExpression, b.parameters(), (block, query) -> {
                     MethodRef md = method(Queryable.TYPE, methodName,
                             functionType(QueryResult.TYPE));
                     Op.Result queryResult = block.op(JavaOp.invoke(queryResultType, md, query));
