+++ test/jdk/java/lang/reflect/code/stream/StreamFuserUsingQuotable.java
@@ -22,6 +22,7 @@
  */
 
 import jdk.incubator.code.*;
+import jdk.incubator.code.analysis.Inliner;
 import jdk.incubator.code.dialect.core.CoreType;
 import jdk.incubator.code.dialect.java.JavaOp;
 import jdk.incubator.code.dialect.java.JavaOp.EnhancedForOp;
@@ -158,11 +159,11 @@ public final class StreamFuserUsingQuotable {
 
             StreamOp sop = streamOps.get(i);
             if (sop instanceof MapStreamOp) {
-                body.inline(sop.op(), List.of(element), (block, value) -> {
+                Inliner.inline(body, sop.op(), List.of(element), (block, value) -> {
                     fuseIntermediateOperation(i + 1, block, value, continueBlock, terminalConsumer);
                 });
             } else if (sop instanceof FilterStreamOp) {
-                body.inline(sop.op(), List.of(element), (block, p) -> {
+                Inliner.inline(body, sop.op(), List.of(element), (block, p) -> {
                     Block.Builder _if = block.block();
                     Block.Builder _else = continueBlock;
                     if (continueBlock == null) {
@@ -175,7 +176,7 @@ public final class StreamFuserUsingQuotable {
                     fuseIntermediateOperation(i + 1, _if, element, _else, terminalConsumer);
                 });
             } else if (sop instanceof FlatMapStreamOp) {
-                body.inline(sop.op(), List.of(element), (block, iterable) -> {
+                Inliner.inline(body, sop.op(), List.of(element), (block, iterable) -> {
                     EnhancedForOp forOp = enhancedFor(block.parentBody(),
                             iterable.type(), ((ClassType) iterable.type()).typeArguments().get(0))
                             .expression(b -> {
@@ -212,7 +213,7 @@ public final class StreamFuserUsingQuotable {
                         Op sourceLoop = loopSupplier.apply(b.parentBody(), source)
                                 .apply(loopBlock -> {
                                     fuseIntermediateOperations(loopBlock, (terminalBlock, resultValue) -> {
-                                        terminalBlock.inline(consumer, List.of(resultValue),
+                                        Inliner.inline(terminalBlock, consumer, List.of(resultValue),
                                                 (_, _) -> {
                                                 });
                                         terminalBlock.op(JavaOp.continue_());
@@ -243,11 +244,11 @@ public final class StreamFuserUsingQuotable {
                     .body(b -> {
                         Value source = b.parameters().get(0);
 
-                        b.inline(supplier, List.of(), (block, collect) -> {
+                        Inliner.inline(b, supplier, List.of(), (block, collect) -> {
                             Op sourceLoop = loopSupplier.apply(block.parentBody(), source)
                                     .apply(loopBlock -> {
                                         fuseIntermediateOperations(loopBlock, (terminalBlock, resultValue) -> {
-                                            terminalBlock.inline(accumulator, List.of(collect, resultValue),
+                                            Inliner.inline(terminalBlock, accumulator, List.of(collect, resultValue),
                                                     (_, _) -> {
                                                     });
                                             terminalBlock.op(JavaOp.continue_());
